pyproject.toml
[tool.poetry]
name = "utils"
version = "0.1.0"
description = "A collection of utility functions and classes"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.57.0"
loguru = "^0.7.0"
pyyaml = "^6.0"
jinja2 = "^3.1.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"